+++ packages/integration_test/test/reporter/data/fail_test_script.dart
@@ -7,12 +7,12 @@ void main() async {
   final IntegrationTestWidgetsFlutterBinding binding =
       IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('passing test', (WidgetTester tester) async {
-    expect(true, true);
+  testWidgets('Failing test 1', (WidgetTester tester) async {
+    expect(false, true);
   });
 
-  testWidgets('failing test', (WidgetTester tester) async {
-    expect(true, false);
+  testWidgets('Failing test 2', (WidgetTester tester) async {
+    expect(false, true);
   });
 
   tearDownAll(() {